requirements.txt
langchain
ollama
pyyaml
langchain-community
